Broke out models into their own files
Instead of one 800+ line monstrocity
