+++ un_project/un_app/models/__init__.py
@@ -1,0 +1,10 @@
+from .player_models import *
+from .building_models import *
+from .company_models import *
+from .denomination_models import *
+from .item_models import *
+from .liability_models import *
+from .liquid_asset_models import *
+from .nation_models import *
+from .territory_models import *
+from .user_profile_models import *
